GrandTheftAutoSanAndreas: WIP

+++ src/GrandTheftAuto3/scenes.ts
@@ -5,11 +5,11 @@ import { GfxDevice, GfxFormat } from '../gfx/platform/GfxPlatform';
 import { DataFetcher } from '../DataFetcher';
 import { GTA3Renderer, SceneRenderer, DrawKey, Texture, TextureArray, MeshInstance, ModelCache, SkyRenderer } from './render';
 import { SceneContext } from '../SceneBase';
-import { getTextDecoder, assert } from '../util';
-import { parseItemPlacement, ItemPlacement, parseItemDefinition, ItemDefinition, ObjectDefinition, ItemInstance, parseZones } from './item';
+import { getTextDecoder } from '../util';
+import { parseItemPlacement, ItemPlacement, parseItemDefinition, ItemDefinition, ObjectDefinition, ItemInstance, parseZones, parseItemPlacementBinary } from './item';
 import { parseTimeCycle, ColorSet } from './time';
 import { parseWaterPro, waterMeshFragData, waterDefinition } from './water';
-import { quat, vec3 } from 'gl-matrix';
+import { vec3 } from 'gl-matrix';
 import { AABB } from '../Geometry';
 import { GfxRendererLayer } from '../gfx/render/GfxRenderer';
 import ArrayBufferSlice from '../ArrayBufferSlice';
@@ -39,7 +39,9 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         },
         ide: [] as string[],
         ipl: [] as string[],
+        ipl_stream: [] as string[],
     };
+    protected versionIMG = 1;
 
     constructor(public id: string, public name: string) {
         this.pathBase = 'GrandTheftAuto3';
@@ -78,20 +80,20 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
     }
 
     private async fetchIMG(dataFetcher: DataFetcher): Promise<void> {
-        if (assetCache.has(`${this.pathBase}/models/gta3.dir`)) return;
-        const [bufferDIR, bufferIMG] = await Promise.all([
-            this.fetch(dataFetcher, 'models/gta3.dir'),
-            this.fetch(dataFetcher, 'models/gta3.img'),
-        ]);
+        if (assetCache.has(`${this.pathBase}/models/gta3.img`)) return;
+        const v1 = (this.versionIMG === 1);
+        const bufferIMG = await this.fetch(dataFetcher, 'models/gta3.img');
+        const bufferDIR = v1 ? await this.fetch(dataFetcher, 'models/gta3.dir') : bufferIMG;
         const view = bufferDIR.createDataView();
-        for (let i = 0; i < view.byteLength; i += 32) {
+        const start = v1 ? 0 : 8;
+        const dirLength = v1 ? view.byteLength : 32 * view.getUint32(4, true);
+        for (let i = start; i < start + dirLength; i += 32) {
             const offset = view.getUint32(i + 0, true);
-            const size = view.getUint32(i + 4, true);
+            const size = v1 ? view.getUint32(i + 4, true) : view.getUint16(i + 4, true);
             const name = UTF8ToString(bufferDIR.subarray(i + 8, 24).createTypedArray(Uint8Array)).toLowerCase();
             const data = bufferIMG.subarray(2048 * offset, 2048 * size);
             assetCache.set(`${this.pathBase}/models/gta3/${name}`, data);
         }
-        assetCache.delete(`${this.pathBase}/models/gta3.img`);
     }
 
     private async fetch(dataFetcher: DataFetcher, path: string): Promise<ArrayBufferSlice> {
@@ -99,7 +101,6 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         let buffer = assetCache.get(path);
         if (buffer === undefined) {
             buffer = await dataFetcher.fetchData(path);
-            assetCache.set(path, buffer);
         }
         return buffer;
     }
@@ -110,19 +111,15 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         return parseItemDefinition(text);
     }
 
-    private async fetchIPL(dataFetcher: DataFetcher, id: string): Promise<ItemPlacement> {
-        if (id === 'test') return {
-            instances: [{
-                id: 0,
-                modelName: 'billboard01',
-                rotation: quat.fromValues(0,0,0,1),
-                translation: vec3.fromValues(0,0,0),
-                scale: vec3.fromValues(10,10,10),
-            }]
-        };
-        const buffer = await this.fetch(dataFetcher, (id === 'props') ? `data/maps/props.IPL` : `data/maps/${id}.ipl`);
-        const text = getTextDecoder('utf8')!.decode(buffer.createDataView());
-        return parseItemPlacement(text);
+    private async fetchIPL(dataFetcher: DataFetcher, id: string, binary = false): Promise<ItemPlacement> {
+        if (binary) {
+            const buffer = await this.fetch(dataFetcher, `models/gta3/${id}.ipl`);
+            return parseItemPlacementBinary(buffer.createDataView());
+        } else {
+            const buffer = await this.fetch(dataFetcher, (id === 'props') ? `data/maps/props.IPL` : `data/maps/${id}.ipl`);
+            const text = getTextDecoder('utf8')!.decode(buffer.createDataView());
+            return parseItemPlacement(text);
+        }
     }
 
     private async fetchTimeCycle(dataFetcher: DataFetcher): Promise<ColorSet[]> {
@@ -150,15 +147,18 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         const buffer = await this.fetch(dataFetcher, txdPath);
         const stream = new rw.StreamMemory(buffer.createTypedArray(Uint8Array));
         const header = new rw.ChunkHeaderInfo(stream);
-        assert(header.type === rw.PluginID.ID_TEXDICTIONARY);
-        const txd = new rw.TexDictionary(stream);
+        if (header.type === rw.PluginID.ID_TEXDICTIONARY) {
+            const txd = new rw.TexDictionary(stream);
+            for (let lnk = txd.textures.begin; !lnk.is(txd.textures.end); lnk = lnk.next) {
+                const texture = new Texture(rw.Texture.fromDict(lnk), txdName, useDXT);
+                cb(texture);
+            }
+            txd.delete();
+        } else {
+            console.error('TXD header type', rw.PluginID[header.type]);
+        }
         header.delete();
         stream.delete();
-        for (let lnk = txd.textures.begin; !lnk.is(txd.textures.end); lnk = lnk.next) {
-            const texture = new Texture(rw.Texture.fromDict(lnk), txdName, useDXT);
-            cb(texture);
-        }
-        txd.delete();
     }
 
     private async fetchDFF(dataFetcher: DataFetcher, modelName: string, cb: (clump: rw.Clump) => void): Promise<void> {
@@ -166,12 +166,15 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         const buffer = await this.fetch(dataFetcher, dffPath);
         const stream = new rw.StreamMemory(buffer.createTypedArray(Uint8Array));
         const header = new rw.ChunkHeaderInfo(stream);
-        assert(header.type === rw.PluginID.ID_CLUMP);
-        const clump = rw.Clump.streamRead(stream);
+        if (header.type === rw.PluginID.ID_CLUMP) {
+            const clump = rw.Clump.streamRead(stream);
+            cb(clump);
+            clump.delete();
+        } else {
+            console.error('DFF header type', rw.PluginID[header.type]);
+        }
         header.delete();
         stream.delete();
-        cb(clump);
-        clump.delete();
     }
 
     protected filter(item: ItemInstance, obj: ObjectDefinition, zone: string) {
@@ -182,6 +185,7 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         await GTA3SceneDesc.initialise();
         const dataFetcher = context.dataFetcher;
         const objects = new Map<string, ObjectDefinition>();
+        const objectIDs = new Map<number, string>();
         const lodnames = new Set<string>();
 
         if (this.complete)
@@ -190,19 +194,21 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         const ides = await Promise.all(this.paths.ide.map(id => this.fetchIDE(dataFetcher, id)));
         for (const ide of ides) for (const obj of ide.objects) {
             objects.set(obj.modelName, obj);
+            if (obj.id !== undefined) objectIDs.set(obj.id, obj.modelName);
             if (obj.modelName.startsWith('lod')) lodnames.add(obj.modelName.substr(3));
         }
         objects.set('water', waterDefinition);
 
-        const ipls = await Promise.all(this.paths.ipl.map(id => this.fetchIPL(dataFetcher, id)));
-        const [colorSets, zones, water] = await Promise.all([this.fetchTimeCycle(dataFetcher), this.fetchZones(dataFetcher), await this.fetchWater(dataFetcher)]);
-        ipls.push(water);
+        const ipls = await Promise.all(
+            this.paths.ipl.map(id => this.fetchIPL(dataFetcher, id)).concat(
+            this.paths.ipl_stream.map(id => this.fetchIPL(dataFetcher, id, true))));
+        const [colorSets, zones] = await Promise.all([this.fetchTimeCycle(dataFetcher), this.fetchZones(dataFetcher)]);
+        //ipls.push(await this.fetchWater(dataFetcher));
 
         const renderer = new GTA3Renderer(device, colorSets, this.weatherTypes, this.water.origin);
-        const loadedTXD = new Map<string, Promise<void>>();
         const loadedDFF = new Map<string, Promise<void>>();
-        const textures  = new Map<string, Texture>();
         const modelCache = new ModelCache();
+        const texturesUsed = new Map<string, Set<string>>();
         const textureSets = new Map<string, Set<Texture>>();
         const drawKeys = new Map<string, DrawKey>();
         const layers = new Map<DrawKey, MeshInstance[]>();
@@ -211,6 +217,13 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         modelCache.meshData.set('water', [waterMeshFragData(this.water.texture)]);
 
         for (const ipl of ipls) for (const item of ipl.instances) {
+            if (item.modelName === undefined && item.id !== undefined) {
+                item.modelName = objectIDs.get(item.id);
+            }
+            if (item.modelName === undefined) {
+                console.error('Missing model name for ID', item.id);
+                continue;
+            }
             const name = item.modelName;
             const haslod = lodnames.has(name.substr(3));
             const obj = objects.get(name);
@@ -229,30 +242,20 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
             }
             if (!this.filter(item, obj, zone)) continue;
 
-            if (!loadedTXD.has(obj.txdName))
-                loadedTXD.set(obj.txdName, this.fetchTXD(device, dataFetcher, obj.txdName, texture => textures.set(texture.name, texture)));
             if (!loadedDFF.has(obj.modelName))
                 loadedDFF.set(obj.modelName, this.fetchDFF(dataFetcher, obj.modelName, clump => modelCache.addModel(clump, obj)));
-            await Promise.all([loadedTXD.get(obj.txdName)!, loadedDFF.get(obj.modelName)!])
+            await loadedDFF.get(obj.modelName)!;
 
-            const model = modelCache.meshData.get(item.modelName);
+            const model = modelCache.meshData.get(name);
             if (model === undefined) {
-                console.warn('Missing model', item.modelName);
+                console.warn('Missing model', name);
                 continue;
             }
-
             for (const frag of model) {
                 if (frag.texName === undefined) continue;
-                const texture = textures.get(frag.texName);
-                if (texture === undefined) {
-                    console.warn('Missing texture', frag.texName, 'for', item.modelName);
-                    continue;
-                }
-
-                let res = '';
-                res += texture.width + 'x' + texture.height + '.' + texture.pixelFormat;
-                if (!textureSets.has(res)) textureSets.set(res, new Set());
-                textureSets.get(res)!.add(texture);
+                const txdName = frag.texName.split('/')[0];
+                if (!texturesUsed.has(txdName)) texturesUsed.set(txdName, new Set());
+                texturesUsed.get(txdName)!.add(frag.texName);
             }
 
             let drawKey = new DrawKey(obj, zone);
@@ -269,10 +272,27 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
         }
 
         const textureArrays = [] as TextureArray[];
+        for (const [txdName, texNames] of texturesUsed) {
+            await this.fetchTXD(device, dataFetcher, txdName, texture => {
+                if (texture.pixels === undefined) return;
+                const texName = texture.name;
+                if (!texNames.has(texName)) return;
+                texNames.delete(texName);
+
+                let res = '';
+                res += texture.width + 'x' + texture.height + '.' + texture.pixelFormat;
+                if (!textureSets.has(res)) textureSets.set(res, new Set());
+                const textureSet = textureSets.get(res)!;
+                textureSet.add(texture);
+                if (textureSet.size >= 0x100) {
+                    textureArrays.push(new TextureArray(device, Array.from(textureSet)));
+                    textureSet.clear();
+                }
+            });
+            if (texNames.size > 0) console.warn('Missing textures', Array.from(texNames), 'from', txdName);
+        }
         for (const [res, textureSet] of textureSets) {
-            const textures = Array.from(textureSet);
-            for (let i = 0; i < textures.length; i += 0x100)
-                textureArrays.push(new TextureArray(device, textures.slice(i, i + 0x100)));
+            textureArrays.push(new TextureArray(device, Array.from(textureSet)));
         }
 
         const sealevel = this.water.origin[2];
@@ -287,10 +307,12 @@ export class GTA3SceneDesc implements Viewer.SceneDesc {
             }
         }
 
+        /*
         await loadedTXD.get('particle')!;
         const waterTex = textures.get(`particle/${this.water.texture}`)!;
         const waterAtlas = new TextureArray(device, [waterTex]);
         renderer.sceneRenderers.push(new SkyRenderer(device, waterAtlas));
+        */
 
         return renderer;
     }
